aloqas_qa_back_fastapi/requirements.txt
fastapi
uvicorn
requests
Pillow
typing
python-multipart
transformers
tensorflow
tf-keras
python-docx
PyMuPDF
chardet
frontend
typing
torch
pydantic
PyPDF2